Update to latest engine version

+++ assets/images/Stages/Level-1/Main-1/Main/Spec/Level-1-Main-1-Main-2Spec.c
@@ -21,7 +21,7 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level_1_Main_1_Main_2Map[];
+extern uint16 Level_1_Main_1_Main_2Map[];
 
 extern CharSetROMSpec Level1Main1MainCharset;
