+++ assets/images/Stages/Level-1/Main-1/Main/Spec/Level-1-Main-1-Main-6Spec.c
@@ -20,7 +20,7 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level_1_Main_1_Main_6Map[];
+extern uint16 Level_1_Main_1_Main_6Map[];
 
 extern CharSetROMSpec Level1Main1MainCharset;
