+++ assets/images/Stages/Level-1/Main-1/MainFront/Spec/Level-1-Main-1-MainFront-2bSpec.c
@@ -19,7 +19,7 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level_1_Main_1_MainFront_2bMap[];
+extern uint16 Level_1_Main_1_MainFront_2bMap[];
 extern CharSetROMSpec Level1Main1MainFrontCharset;
