+++ assets/images/Stages/Level-1/Main-1/MainFront/Spec/Level-1-Main-1-MainFront-3bSpec.c
@@ -18,7 +18,7 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level_1_Main_1_MainFront_3bMap[];
+extern uint16 Level_1_Main_1_MainFront_3bMap[];
 extern CharSetROMSpec Level1Main1MainFrontCharset;
